lots of minor changes - mostly explicit type casting - to shut up Xcode's warnings when compiling Tap.Tools 2.2.
git-svn-id: http://ttblue.googlecode.com/svn/trunk@48 bc2ad3ca-9a4d-0410-8d14-35da9c29b421

+++ ttblue/dsplib/tt_comb.h
@@ -65,7 +65,7 @@ class tt_comb:public tt_audio_base{
 			buffersize = arg > 24 ? arg : 25;		// init the size of the delay buffer - minimum is 25
 		
 			// ALLOCATE MEMORY BUFFER
-			buffersize_in_samples = buffersize * (sr * 0.001);
+			buffersize_in_samples = (long int)(buffersize * (sr * 0.001));
 			if (buffersize_in_samples < 24)			// minimum delay buffer size is 24 samples
 					buffersize_in_samples = 24;
 			alloc = (float *)mem_alloc((buffersize_in_samples + 4) * sizeof(float));
